+++ CHANGELOG.md
@@ -630,3 +630,4 @@
 - Added Refresh Rate slider to settings UI
 - Fix resolution dropdown showing incorrectly if first resolution is chosen
 - MappedInputModule now updates scroll speed of MappedInputComponent on config change
+- Reworked difficulty system, making DifficultyValues a partial class for extensibility and loading from a JSON file in RPGDefs